rename websocket task to serverless task
